feat(streaming): implement robust streaming optimization layer
- Add StreamingRepository with throughput smoothing and transport preferences
- Implement CDN domain classification with CdnMatcher utility
- Add TrafficRepository integration for throughput forwarding
- Integrate binder reconnect handling in TProxyService
- Enhance outbound priority chain with QUIC/H2 optimizations
- Add StreamingPanel UI component (behind feature flag)
- Implement DNS/sniff race hardening in RouteLookupEngine
- Wire invalidation hooks for network/config/screen-off events
- Add BuildConfig feature flag for streaming panel

Features:
- Sliding-window throughput smoothing (window=5, drop max/min)
- Transport preference cache with 60s TTL (QUIC if RTT < 110ms)
- CDN classification for major streaming domains
- Bitrate drop detection and stall mitigation
- Binder reconnect with automatic callback re-registration
- Outbound priority tagging with BBR, keepalive, reusePort

+++ app/src/main/jni/perf-net/src/perf_crypto_neon.cpp
@@ -250,7 +250,7 @@ Java_com_simplexray_an_performance_PerformanceManager_nativeAES128Encrypt(
     // Validate key length (16 bytes for AES-128)
     jlong key_capacity = env->GetDirectBufferCapacity(key);
     if (key_capacity < 16) {
-        LOGE("Invalid key length: %ld (required: 16)", key_capacity);
+        LOGE("Invalid key length: %lld (required: 16)", (long long)key_capacity);
         return -1;
     }
     
@@ -474,14 +474,14 @@ Java_com_simplexray_an_performance_PerformanceManager_nativeChaCha20NEON(
     // Validate key length (32 bytes for ChaCha20)
     jlong key_capacity = env->GetDirectBufferCapacity(key);
     if (key_capacity < 32) {
-        LOGE("Invalid key length: %ld (required: 32)", key_capacity);
+        LOGE("Invalid key length: %lld (required: 32)", (long long)key_capacity);
         return -1;
     }
     
     // Validate nonce length (12 bytes for ChaCha20)
     jlong nonce_capacity = env->GetDirectBufferCapacity(nonce);
     if (nonce_capacity < 12) {
-        LOGE("Invalid nonce length: %ld (required: 12)", nonce_capacity);
+        LOGE("Invalid nonce length: %lld (required: 12)", (long long)nonce_capacity);
         return -1;
     }
